[SYCL][ESIMD] Split test into core and fp_extra files
